+++ TractorMod/release-notes.md
@@ -7,6 +7,7 @@
 * Simplified the `config.json` by replacing per-tool settings with flags and common settings.
 * Can no longer summon tractor before building a tractor garage.
 * Fixed tractor speed debuff lasting a second after you dismount tractor.
+* Fixed hoe destroying objects.
 
 ## 3.2.1
 * Fixed initialisation errors.